Fix process container build
